Fix crypt_blowfish lib lookup
@*INC doesn't contain just a list of paths anymore, it is currently a list
of file# or inst# strings (which are strings of the actual paths, however
with file# or inst# in front; I believe these prefixes are to denote a local
file system location or an installation location).  Thus these paths can't
be found on the filesystem.  The hack used here is to strip the leading
'file#' or 'inst#' strings and use the remaining path to perform the library
lookup in a similar manner to that currently done.  This change makes the
test suite run again.