Handle file handle closing when main thread exits
END is called when the main thread terminates, even if other threads
are still running. This causes a potential attempt to read from a
closed filehandle. Using a callable varialbe to handle this seems to
be the best way to do this, to avoid needless checks everytime a salt
is generated.